% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{check if a number is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
returns a logical vector that indicates wheter the numbers in x are odd or not
}
\description{
A short description ou on montre \code{x} comme un nombre
}
\examples{
is_odd(1)
is_odd(2)


}
